+++ packages/cli/src/config/index.js
@@ -1,0 +1,3 @@
+const { loadConfig } = require('./load')
+
+module.exports = { loadConfig }