+++ packages/cli/src/tasks/index.js
@@ -1,0 +1,4 @@
+const { importSources, extract } = require('./import-sources')
+const { startServer } = require('./server')
+
+module.exports = { importSources, extract, startServer }